internal/conf: Move GetDeduplicatedForksIndex to compute.go
Also move the related tests.
