+++ internal/conf/computed_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	licensing "github.com/sourcegraph/sourcegraph/internal/accesstoken"
+	"github.com/sourcegraph/sourcegraph/internal/collections"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/conf/deploy"
 	"github.com/sourcegraph/sourcegraph/lib/pointers"
@@ -309,6 +310,53 @@ func TestCodyEnabled(t *testing.T) {
 			Mock(&Unified{SiteConfiguration: test.sc})
 			have := CodyEnabled()
 			assert.Equal(t, test.want, have)
+		})
+	}
+}
+
+func TestGetDeduplicatedForksIndex(t *testing.T) {
+	testCases := []struct {
+		name       string
+		haveConfig *schema.Repositories
+		wantIndex  collections.Set[string]
+	}{
+		{
+			name:      "config not set",
+			wantIndex: map[string]struct{}{},
+		},
+		{
+			name:       "repositories set, but deduplicated forks is empty",
+			haveConfig: &schema.Repositories{},
+			wantIndex:  map[string]struct{}{},
+		},
+		{
+			name: "deduplicated forks is not empty",
+			haveConfig: &schema.Repositories{
+				DeduplicateForks: []string{
+					"abc",
+					"def",
+					"abc", // a duplicate
+				},
+			},
+			wantIndex: map[string]struct{}{
+				"abc": {},
+				"def": {},
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			Mock(&Unified{
+				SiteConfiguration: schema.SiteConfiguration{
+					Repositories: tc.haveConfig,
+				},
+			})
+
+			gotIndex := GetDeduplicatedForksIndex()
+			if diff := cmp.Diff(gotIndex, tc.wantIndex); diff != "" {
+				t.Errorf("mismatched deduplicated repos index: (-want, +got)\n%s", diff)
+			}
 		})
 	}
 }
